pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "arxiv-diff"
version = "0.1.0"
description = "AI agent that generates human-readable changelogs between versions of arXiv papers"
authors = [{name = "arxiv-diff", email = "contact@arxiv-diff.dev"}]
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "pymupdf>=1.24.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "typer>=0.12.0",
    "python-dotenv",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "sqlalchemy>=2.0.0",
    "ollama>=0.3.0"
]

[project.scripts]
arxiv-diff = "arxiv_diff.main:app"
